Config: make every percentage option a percentage, and reject wrong units
gamma was the odd one out. luminance, saturation, volume and balance are whole
percentages that the bridges divide by 100; gamma was a raw 1.0-2.0 exponent
passed straight through, so a caller who reasonably assumed the unit of the two
options either side of it got pow(channel, 100) and an almost entirely black
screen (ares screen.cpp:387-391). Fixing that only in the demo would leave every
other consumer — and every agent reading the signature — to fall in the same
hole, so the API changes instead: gamma is an int percentage, 50-200, 100
unchanged.

Out-of-range values are refused with INVALID_PARAMETERS naming the unit rather
than clamped. That is deliberate migration UX: code still passing 1.0 is told
what the unit is instead of rendering black. Both bridges share a percent()
helper so the platforms cannot drift on it.

speed stays a 0.25-4.0 multiplier and now says so — 1.0 = native speed is too
universal a convention to bend for symmetry. Every other numeric option's unit
is spelled out on Config for the same reason: the signature should answer the
question without a trip to the native source.

Two more things the audit turned up:

- iOS SetAudio defaulted omitted knobs to ?? 100 / ?? 0, so setting balance
  alone silently reset volume to full. It now merges like Android's, holding the
  current values on the renderer.
- applyPresentation() batched both audio knobs to work around exactly that, and
  its comment stated the reset as the contract. With the platforms agreed, the
  claim was false; the batching is harmless so only the comment goes.

Verified on the Thor: 86/86 conformance steps against the real native layer,
including new steps that reject a multiplier-style gamma, accept a percentage
one, and reject an out-of-range volume.

+++ src/Config/Config.php
@@ -16,6 +16,17 @@ use KevinBatdorf\RetroEmulator\VideoOutput;
 class Config
 {
     /**
+     * Every percentage option below is a whole percent where 100 means "leave it
+     * alone", so one mental model covers the lot. `speed` is the exception and
+     * says so: it is a multiplier, because 1.0 = native speed is universal.
+     *
+     * @param  int|null  $luminance  Percent, 0–100; 100 is untouched.
+     * @param  int|null  $saturation  Percent, 0–100; 100 is untouched.
+     * @param  int|null  $gamma  Percent, 50–200; 100 is untouched. Above 100
+     *                           darkens midtones, below brightens them.
+     * @param  int|null  $volume  Percent, 0–100.
+     * @param  int|null  $balance  Percent, −100 (full left) … +100 (full right).
+     * @param  float|null  $speed  Multiplier, 0.25–4.0; 1.0 is native speed.
      * @param  bool|null  $overscan  Trims the overscan border (trims by default);
      *                               no effect on systems with no overscan region.
      * @param  bool|null  $colorBleed  Composite color-bleed filter; no effect on
@@ -23,13 +34,14 @@ class Config
      * @param  InputCapture|null  $inputCapture  Resolved when the surface is
      *                                           created; not changeable at runtime.
      * @param  int|null  $runAhead  0 or 1 — ares runs exactly one hidden frame.
+     * @param  int|null  $rewindBufferSeconds  Seconds of history to retain.
      * @param  string|null  $shader  librashader .slangp path; clear an active
      *                               shader with setShader(null), not null here.
      */
     public function __construct(
         public ?int $luminance = null,
         public ?int $saturation = null,
-        public ?float $gamma = null,
+        public ?int $gamma = null,
         public ?bool $colorBleed = null,
         public ?bool $overscan = null,
         public ?VideoOutput $output = null,
