+++ src/Config/RegionalSystemConfig.php
@@ -18,7 +18,7 @@ abstract class RegionalSystemConfig extends SystemConfig
     public function __construct(
         ?int $luminance = null,
         ?int $saturation = null,
-        ?float $gamma = null,
+        ?int $gamma = null,
         ?bool $colorBleed = null,
         ?bool $overscan = null,
         ?VideoOutput $output = null,
