+++ src/Config/SystemConfig.php
@@ -16,7 +16,7 @@ abstract class SystemConfig extends Config
     public function __construct(
         ?int $luminance = null,
         ?int $saturation = null,
-        ?float $gamma = null,
+        ?int $gamma = null,
         ?bool $colorBleed = null,
         ?bool $overscan = null,
         ?VideoOutput $output = null,
